Reset Syndication src\ (#25059)
* Reset Syndication \src

* Marked a test with ActiveIssue.

* CodeFormatted Syndication src\.

* Removed old header.

* Minor changes.

* Remove unused string resources

* Second round of code formatting

+++ src/System.ServiceModel.Syndication/tests/BasicScenarioTests.cs
@@ -222,13 +222,13 @@ namespace System.ServiceModel.Syndication.Tests
                 File.Delete(AtomPath);
             }
         }
-        
+
         [Fact]
         public static void SyndicationFeed_Load_Rss()
         {
             XmlReaderSettings setting = new XmlReaderSettings();
             using (XmlReader reader = XmlReader.Create(@"rssSpecExample.xml", setting))
-            { 
+            {
                 SyndicationFeed rss = SyndicationFeed.Load(reader);
                 Assert.True(rss.Items != null);
             }
@@ -246,6 +246,7 @@ namespace System.ServiceModel.Syndication.Tests
         }
 
         [Fact]
+        [ActiveIssue(25080)]
         public static void SyndicationFeed_Rss_TestDisjointItems()
         {
             using (XmlReader reader = XmlReader.Create(@"RssDisjointItems.xml"))
@@ -266,6 +267,7 @@ namespace System.ServiceModel.Syndication.Tests
 
 
         [Fact]
+        [ActiveIssue(25080)]
         public static void SyndicationFeed_Atom_TestDisjointItems()
         {
             using (XmlReader reader = XmlReader.Create(@"AtomDisjointItems.xml"))
@@ -285,6 +287,7 @@ namespace System.ServiceModel.Syndication.Tests
         }
 
         [Fact]
+        [ActiveIssue(25080)]
         public static void SyndicationFeed_Rss_WrongDateFormat()
         {
             // *** SETUP *** \\
@@ -365,6 +368,14 @@ namespace System.ServiceModel.Syndication.Tests
         }
 
         [Fact]
+        [ActiveIssue(25058)]
+        public static void AtomFeedPositiveTest_WsFeedIcon()
+        {
+            string file = "ws-feed-icon.xml";
+            AtomFeedPositiveTest(file);
+        }
+
+        [Fact]
         public static void AtomFeedPositiveTest()
         {
             string dataFile = @"atom_feeds.dat";
@@ -372,45 +383,57 @@ namespace System.ServiceModel.Syndication.Tests
 
             foreach (string file in fileList)
             {
-                string serializeFilePath = Path.GetTempFileName();
-                try
+                AtomFeedPositiveTest(file);
+            }
+        }
+
+        private static void AtomFeedPositiveTest(string file)
+        {
+            string serializeFilePath = Path.GetTempFileName();
+            bool toDeletedFile = true;
+            try
+            {
+                SyndicationFeed feedObjct;
+
+                using (XmlReader reader = XmlReader.Create(file))
                 {
-                    SyndicationFeed feedObjct;
-
-                    using (XmlReader reader = XmlReader.Create(file))
-                    {
-                        feedObjct = SyndicationFeed.Load(reader);
-                        reader.Close();
-                    }
-
-                    using (XmlWriter writer = XmlWriter.Create(serializeFilePath))
-                    {
-                        Atom10FeedFormatter f = new Atom10FeedFormatter(feedObjct);
-                        f.WriteTo(writer);
-                        writer.Close();
-                    }
-
-                    CompareHelper ch = new CompareHelper
-                    {
-                        Diff = new XmlDiff()
-                        {
-                            Option = XmlDiffOption.IgnoreComments | XmlDiffOption.IgnorePrefix | XmlDiffOption.IgnoreWhitespace | XmlDiffOption.IgnoreChildOrder | XmlDiffOption.IgnoreAttributeOrder
-                        },
-                        AllowableDifferences = GetAtomFeedPositiveTestAllowableDifferences()
-                    };
-                    Assert.True(ch.Compare(file, serializeFilePath), $"Failed File Name:{file}");
+                    feedObjct = SyndicationFeed.Load(reader);
+                    reader.Close();
                 }
-                catch (Exception e)
+
+                using (XmlWriter writer = XmlWriter.Create(serializeFilePath))
                 {
-                    Console.WriteLine();
-                    Console.WriteLine("------------------------------");
-                    Console.WriteLine($"Failed File Name:{file}");
-                    throw e;
+                    Atom10FeedFormatter f = new Atom10FeedFormatter(feedObjct);
+                    f.WriteTo(writer);
+                    writer.Close();
                 }
-                finally
+
+                CompareHelper ch = new CompareHelper
                 {
+                    Diff = new XmlDiff()
+                    {
+                        Option = XmlDiffOption.IgnoreComments | XmlDiffOption.IgnorePrefix | XmlDiffOption.IgnoreWhitespace | XmlDiffOption.IgnoreChildOrder | XmlDiffOption.IgnoreAttributeOrder
+                    },
+                    AllowableDifferences = GetAtomFeedPositiveTestAllowableDifferences()
+                };
+
+                if (!ch.Compare(file, serializeFilePath))
+                {
+                    toDeletedFile = false;
+                    Assert.True(false, $"The generated file was different from the baseline file{Environment.NewLine}Baseline: {file}{Environment.NewLine}Actual: {serializeFilePath}");
+                }
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine();
+                Console.WriteLine("------------------------------");
+                Console.WriteLine($"Failed File Name:{file}");
+                throw e;
+            }
+            finally
+            {
+                if (toDeletedFile)
                     File.Delete(serializeFilePath);
-                }
             }
         }
 
@@ -481,7 +504,7 @@ namespace System.ServiceModel.Syndication.Tests
             string file;
             using (StreamReader sr = new StreamReader(dataFile))
             {
-                while(!string.IsNullOrEmpty(file = sr.ReadLine()))
+                while (!string.IsNullOrEmpty(file = sr.ReadLine()))
                 {
                     if (!file.StartsWith("#"))
                     {
